refactor: test 코드 폴더 구조를 src 폴더 구조와 동일하게 수정
