+++ __tests__/Model/LottoGame.test.js
@@ -1,6 +1,6 @@
-import LottoGame from "../src/Model/LottoGame.js";
-import LottoList from "../src/Model/LottoList.js";
-import Random from "../src/utils/Random.js";
+import LottoGame from "../../src/Model/LottoGame.js";
+import LottoList from "../../src/Model/LottoList.js";
+import Random from "../../src/utils/Random.js";
 
 const mockRandoms = (numbers) => {
   Random.randomArray = jest.fn();
