+++ __tests__/utils/parsing.test.js
@@ -1,5 +1,5 @@
-import { ERROR_MESSAGE } from "../src/constants/errorMessage.js";
-import { parsingNumbers } from "../src/utils/parsing.js";
+import { ERROR_MESSAGE } from "../../src/constants/errorMessage.js";
+import { parsingNumbers } from "../../src/utils/parsing.js";
 
 describe("문자열 파싱 테스트", () => {
   test("로또 번호 파싱 시의 유효성 검사", () => {
